+++ Oracle11gBase9787302458227/chapter04/ch040302_Where.py
@@ -1,6 +1,7 @@
 """ 章节4.3.3 分组查询 """
 
 import cx_Oracle as cx
+from Oracle11gBase9787302458227.chapter04 import OptOracle
 
 # 连接 Oracle 数据库
 con_scott = cx.connect('scott', 'tiger', '127.0.0.1:1521/orcl')  # SCOTT 模式
@@ -77,18 +78,4 @@ tableHead4_34 = ['empno', 'ename', 'sal']
 sqlCase4_34 = "select empno, ename, sal from emp where sal < 2000 or sal > 3000"
 
 
-# 连接数据库，执行 SELECT 语句，打印数据
-def printData(connect, tableHead, sqlCase):
-    cursor = connect.cursor()  # 创建游标
-    print(tableHead)  # 打印表头
-    cursor.execute(sqlCase)  # 执行sql语句，可以替换不同的例子（93710）
-
-    # 遍历游标，打印数据
-    for result in cursor:
-        print(result)
-
-    cursor.close()  # 关闭游标
-    connect.close()  # 关闭数据库连接
-
-
-printData(con_scott, tableHead4_34, sqlCase4_34)  # 执行 exp4_34
+OptOracle.printData(con_scott, tableHead4_34, sqlCase4_34)  # 执行 exp4_34
